# src/main/java/com/backend/api/IndividualsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.backend.api;

import com.backend.model.Individual;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-03T01:08:01.089826924Z[Etc/UTC]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "individuals", description = "the individuals API")
public interface IndividualsApi {

    /**
     * GET /individuals : 個体情報一覧を取得
     *
     * @return 一覧取得成功 (status code 200)
     */
    @Operation(
        operationId = "individualsGet",
        summary = "個体情報一覧を取得",
        responses = {
            @ApiResponse(responseCode = "200", description = "一覧取得成功", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Individual.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/individuals",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<Individual>> individualsGet(
        
    );


    /**
     * POST /individuals : 個体情報を新規登録
     *
     * @param individual  (required)
     * @return 登録成功 (status code 201)
     */
    @Operation(
        operationId = "individualsPost",
        summary = "個体情報を新規登録",
        responses = {
            @ApiResponse(responseCode = "201", description = "登録成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Individual.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/individuals",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Individual> individualsPost(
        @Parameter(name = "Individual", description = "", required = true) @Valid @RequestBody Individual individual
    );


    /**
     * DELETE /individuals/{species_cd}/{id} : 個体情報を論理削除
     *
     * @param speciesCd  (required)
     * @param id  (required)
     * @return 削除成功 (status code 204)
     */
    @Operation(
        operationId = "individualsSpeciesCdIdDelete",
        summary = "個体情報を論理削除",
        responses = {
            @ApiResponse(responseCode = "204", description = "削除成功")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/individuals/{species_cd}/{id}"
    )
    
    ResponseEntity<Void> individualsSpeciesCdIdDelete(
        @Parameter(name = "species_cd", description = "", required = true, in = ParameterIn.PATH) @PathVariable("species_cd") String speciesCd,
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * PUT /individuals/{species_cd}/{id} : 個体情報を更新
     *
     * @param speciesCd  (required)
     * @param id  (required)
     * @param individual  (required)
     * @return 更新成功 (status code 200)
     */
    @Operation(
        operationId = "individualsSpeciesCdIdPut",
        summary = "個体情報を更新",
        responses = {
            @ApiResponse(responseCode = "200", description = "更新成功")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/individuals/{species_cd}/{id}",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> individualsSpeciesCdIdPut(
        @Parameter(name = "species_cd", description = "", required = true, in = ParameterIn.PATH) @PathVariable("species_cd") String speciesCd,
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "Individual", description = "", required = true) @Valid @RequestBody Individual individual
    );

}